% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForHasse.R
\name{compare.profiles}
\alias{compare.profiles}
\title{Compare each pair of structuples/profiles in a matrix.}
\usage{
compare.profiles(pr)
}
\arguments{
\item{pr}{A matrix where each row is a distinct structuple/profile.}
}
\value{
A logical square matrix with 'true' indicating less than or equal relation for input to HasseDiagram-Hasse function.
}
\description{
Compare each pair of structuples/profiles in a matrix.
}
\examples{
CompPairMat=compare.profiles(as.matrix(SimplePOSACdata[,1:4]))
hasseDiagram::hasse(CompPairMat)

CRASCompMat=compare.profiles(as.matrix(CRASdata[,1:5]))
#the result of this function can be used to make a Hasse Diagram
hasseDiagram::hasse(CRASCompMat)
}
