% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRASdata.R
\docType{data}
\name{CRASdata}
\alias{CRASdata}
\title{A data set of CRAS demand patterns.}
\format{A data frame with 32 rows and 5 variables:
\describe{
  \item{complex}{Rating of item complexity from 1-5}
  \item{resource}{Rating of required resources from 1-5}
  \item{abstract}{Rating of item abstractness from 1-5}
  \item{taskstr}{Rating of task strategy from 1-5}
  \item{respstr}{Rating of response strategy from 1-5}
  \item{Freq}{Frequency of occurence of each pattern}
}}
\usage{
CRASdata
}
\description{
A data set of 32 patterns of CRAS demands of items from a Physics examination.
}
\keyword{datasets}
