% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamfunc.R
\name{gamfunc}
\alias{gamfunc}
\title{Function to create the gamma matrix}
\usage{
gamfunc(CR)
}
\arguments{
\item{CR}{A matrix of ratings. Rows should represent observations and columns should represent variables.}
}
\description{
This function calculates the gamma matrix. That is, the matrix indicating whether each unique pattern is superior (+1), inferior (-1) or
incomparable (0) to each other unique pattern.
}
\examples{
gamfunc(SimplePOSACdata[,1:4])
}
