% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForHasse.R
\name{\%gte\%}
\alias{\%gte\%}
\title{Function to compare 2 structuples/profiles for 'greater than or equal' relation}
\usage{
x \%gte\% y
}
\arguments{
\item{x}{First structuple/profile}

\item{y}{Second structuple/profile}
}
\description{
Function to compare 2 structuples/profiles for 'greater than or equal' relation
}
\examples{
c(1,2)\%gte\%c(2,1)
c(1,2,1,5)\%gte\%c(1,2,1,2)
}
