% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForHasse.R
\name{\%lte\%}
\alias{\%lte\%}
\title{Function to compare 2 structuples/profiles for 'less than or equal' relation}
\usage{
x \%lte\% y
}
\arguments{
\item{x}{First structuple/profile}

\item{y}{Second structuple/profile}
}
\description{
Function to compare 2 structuples/profiles for 'less than or equal' relation
}
\examples{
c(1,2)\%lte\%c(2,1)
c(1,2,1,5)\%lte\%c(11,2,1,7)
}
