% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemdiagram.R
\name{itemdiagram}
\alias{itemdiagram}
\title{Function to plot X and Y output from POSAC with colours for different values of given variables}
\usage{
itemdiagram(posacout, vars)
}
\arguments{
\item{posacout}{Output from function POSAC (or POSACupdate).}

\item{vars}{Indices of items from pattern matrix. Including too many items at once may possibly produce an error. Also note that if multiple items are specified, the positioning of points will be the same in each plot generated - only the colouring and labelling will change.}
}
\description{
This function produces plots of the values of the two variables (X and Y) produced by the analysis (e.g. the POSAC function) against one another. Colour is added to the chart to represent the values of a particular individual variable within the data set. 
These plots may help to identify whether a given indidvidual variable is particularly important in the definition of either X or Y.
Note that if multiple plots are produced (for multiple items) the positions of points will be identical in every plot. 
Only the colouring and labelling of points will vary.
}
\examples{
posac2=POSAC(CRASdata[,1:5],CRASdata[,6])
itemdiagram(posac2,1)
itemdiagram(posac2,2)
itemdiagram(posac2,1:3)
itemdiagram(posac2,1:5)
}
\keyword{Scalogram}
