% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monotonicity.R
\name{mu}
\alias{mu}
\title{Function to calculate Guttman's monotonicity coefficient (mu) between two vectors}
\usage{
mu(x, y)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\value{
a number between -1 and +1
}
\description{
The coefficient expresses the extent to which values on one variable increase in a particular direction along with the values on
another variable without assuming that the increase is exactly according to a straight line.
}
\examples{
x1=rnorm(1000)
x2=x1^3+rnorm(1000)
#Compare three coefficients of association and plot association
cor(x1,x2)
cor(x1,x2,method="spearman")
mu(x1,x2)
plot(x1,x2)
}
