% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSAC.R
\name{POSAC}
\alias{POSAC}
\title{Partial Order Scalogram Analysis with Base Coordinates.}
\usage{
POSAC(patmat, freqs)
}
\arguments{
\item{patmat}{A matrix of patterns of values across. Each row of the matrix should represents a distinct pattern with no duplicates. The columns are the variables defining the patterns.}

\item{freqs}{A vector of frequencies of each pattern. This should have the same length as nrow(patmat).}
}
\value{
The function returns a list with the following elements:
\describe{
  \item{Criteria}{The output from the function realcriteria for the final solution.}
  \item{X}{The ranking of each pattern on the first dimension of the returned POSAC solution.}
  \item{Y}{The ranking of each pattern on the second dimension of the returned POSAC solution.}
  \item{Patterns}{A character vector of concatenated patterns.}
  \item{patmat}{The pattern matrix used in analysis to begin with.}
  \item{freqs}{The frequency with which in pattern occurs.}
  \item{tidyframe}{A single data frame containing the structuples(patterns), frequencies, X, Y, rescaled X and Y to be between a 0 and 100, J (=rescaled X+rescaled Y)and L(=100+rescaled X-rescaled Y).}
}
}
\description{
This functions takes a matrix of patterns and the frequencies of the patterns and performs POSAC analysis.
}
\examples{
#POSAC analysis of some simple patterns
posac1=POSAC(SimplePOSACdata[,1:4],SimplePOSACdata[,5])
posac1
plot(posac1$X,posac1$Y,type='n',xlab="Facet analysis X",ylab="Facet analysis Y")
text(posac1$X,posac1$Y,posac1$Patterns)

#POSAC analysis of some CRAS patterns
posac2=POSAC(CRASdata[,1:5],CRASdata[,6])
posac2
plot(posac2$X,posac2$Y,type='n',xlab="Facet analysis X",ylab="Facet analysis Y")
text(posac2$X,posac2$Y,posac2$Patterns)
posac2$tidyframe

#An enhanced plot with colour used to indicate the frequency of patterns
library(ggplot2)
g1=ggplot(data=data.frame(X=posac2$X
				,Y=posac2$Y
				,pat=posac2$Patterns
				,Freq=posac2$freqs),
	aes(x=X,y=Y,label=pat,col=Freq))
g1+geom_text()+theme_minimal()+labs(x="Facet analysis X",y="Facet analysis Y")

#make an item diagram using ggplot2 (can also be done using the itemdiagram function)
g2=ggplot(data=data.frame(X=posac2$X,Y=posac2$Y),
          aes(x=X,y=Y,label=posac2$patmat$complex,col=as.factor(posac2$patmat$complex)))
g2+geom_text()+geom_point()+theme_minimal()+labs(x="Facet analysis X",y="Facet analysis Y")+guides(col="none")
}
\keyword{Scalogram}
