% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realcriteria.R
\name{realcriteria}
\alias{realcriteria}
\title{Properly calculated POSAC criteria.}
\usage{
realcriteria(XY, gam, Freq, printwarn = TRUE)
}
\arguments{
\item{XY}{A concatenated vector of the the x-coordinates and y-coordinates for each combination of values for the covariates (i.e. should have length 2n if n if the number of combinations we are interested in).}

\item{gam}{The matrix of gamma values.}

\item{Freq}{The number of observations with each combination of covariate values.}

\item{printwarn}{Logical input denoting whether the function should print a message about the number of equal pairs not counted in calculations.}
}
\description{
This function counts (weighted) number of pairs that are in the correct order in terms of X and Y (realU merely calculates whether differences have been labelled as comparable or incomparable correctly).
}
\examples{
posac2=POSAC(CRASdata[,1:5],CRASdata[,6])
realcriteria(c(posac2$X,posac2$Y),gamfunc(CRASdata[,1:5]),CRASdata[,6])
#Compare to applying random X and Y coordinates
realcriteria(c(rnorm(dim(CRASdata)[1]),rnorm(dim(CRASdata)[1])),gamfunc(CRASdata[,1:5]),CRASdata[,6],printwarn=FALSE)
}
