% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realU.R
\name{realU}
\alias{realU}
\title{Function to calculate the U matrix.}
\usage{
realU(XY, gam2, Freq)
}
\arguments{
\item{XY}{A concatenated vector of the the x-coordinates and y-coordinates for each combination of values for the covariates (i.e. should have length 2n if n if the number of combinations we are interested in).}

\item{gam2}{The matrix of squared gamma values (gam2)}

\item{Freq}{The number of observations with each combination of covariate values.}
}
\description{
This function is not intended for use directly but rather is used as a subroutine of the main POSAC function.
It counts the (weighted) number of pairs that are correctly positioned as either comparable or non-comparable.
}
