% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resortPOSAC.R
\name{resortPOSAC}
\alias{resortPOSAC}
\title{Function to swap initial results so that greater and lesser are distringuished.}
\usage{
resortPOSAC(XY, gam)
}
\arguments{
\item{XY}{A concatenated vector of the the x-coordinates and y-coordinates for each combination of values for the covariates (i.e. should have length 2n if n if the number of combinations we are interested in).}

\item{gam2}{The matrix of squared gamma values (gam2)}
}
\description{
Sorts starting with lowest values of J and working to highest. This function is not intended for use directly but rather is used as a subroutine of the main POSAC function.
}
