% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resortPOSAC2.R
\name{resortPOSAC2}
\alias{resortPOSAC2}
\title{Further optimisation via sorting.}
\usage{
resortPOSAC2(XY, gam, CR)
}
\arguments{
\item{XY}{A concatenated vector of the the x-coordinates and y-coordinates for each combination of values for the covariates (i.e. should have length 2n if n if the number of combinations we are interested in).}

\item{CR}{The matrix of ratings.}

\item{gam2}{The matrix of squared gamma values (gam2)}
}
\description{
This function is intended to ensure broad directions are corrected (i.e. most dominant in top right and least dominant in bottom left) and then
performs sorting to improve the target criteria.
This function is not intended for use directly but rather is used as a subroutine of the main POSAC function.
}
