% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplePOSACdata.R
\docType{data}
\name{SimplePOSACdata}
\alias{SimplePOSACdata}
\title{A data set of 11 patterns from four variables.}
\format{A data frame with 11 rows and 5 variables:
\describe{
  \item{f1-f4}{Variables with values of either 1 or 2}
  \item{Freq}{Frequency of occurence of each pattern}
}}
\usage{
SimplePOSACdata
}
\description{
A simple data set of 11 different patterns of four variables and the frequency of each pattern.
}
\keyword{datasets}
